JCache backup operation do not propagate CacheNotExistsException
Reasoning:
1. I am a partition owner. You are a backup replica. I execute PUT and send you BACKUP
2. before you process the BACKUP someone else will tell you to destroy the cache
3. when you start processing the BACKUP then the cache no longer exists -> exception

New cache creation is blocking a caller until the cache is created on all members
->
if backup operation does not find the cache then it's because it was already destroyed.
not because it wasn't created yet.
->
Skipping the backup when a cache is not found will not lose any item.

(cherry picked from commit 99bd335)