Simplifies sending multiple backups
In case of sending multiple backups, the backupoperation (so the one doing the real
backup work) is converted to data and this is given to the Backup. This probably is
done to prevent sharing the same operation instance.

However Backup instances are always send using the OperationService.send(Operation,target)
and this method always converts to Data. So there is no need to clone to backupoperation;
it can safely be shared between multiple backup instances since the backups are first
serialized/deserialized before the actual backup operation is used. So always a cloned
instance is used.