Make client-side map queries tolerant to cluster migrations
While invoking queries on members, it is possible that the
query might fail during migrations with exceptions like
TargetNotMemberException, MemberLeftException etc. These
exceptions are normal during migrations and should not be
propagated back to the user, since in any case we will be able
to recover by querying the specific partitions with
QueryPartitionOperations. This aligns the behaviour of the
client-side map proxy with the member-side.