Adds toObject(Data, Class) interfce to SerializationService. This interface allows deserialization using a known object type.
Adds the capability to deserialize a DataSerializable using a custom class name in the derialized data rather than the actual object class name (e.g. the derived class name).

Adds client and server version exchange during connection authentication.

Adds compatibility fixes for Cache createconfig and getconfig requests and responses.

Adds a 3.6.x compatibility switch to the server and client. Namely, the following properties: "hazelcast.compatibility.3.6.client" , "hazelcast.compatibility.3.6.server"  . By default, this property is not set and 3.7.3, 3.8-SNAPSHOT will work with 3.7.x but not with 3.6.x. To make it work with 3.6.x, then the switch should be used.