Repair partition table during join
Replace old and new addresses of a rolling restarting node in partition table
during join. When Hot Restart is enabled, even address of a node changes,
its UUID will remain intact. So, it an existing node restarts with a new address,
its old and new addresses are replaced in partition table.