Fixes the client listener registration if the client owner connection is lost and reconnected (e.g. due to heartbeat or network problems).
Authentication response returns a list of members at which the client is unregistered so that the listener service acts upon this list. The smart listener service re-registers the listeners to these unregistered members. The non-smart listener service re-registers the listeners only if the reconnected member is not the same as the previous member or the previous member client resources were cleaned up.

The ClientDisconnectionOperation also acts smart so that it only cleans client resources if the dead member uuid is the current owner member uuid of the client and there is no alive connection endpoint to the client.

For compatibility, when client is connected to earlier versions the smart listener service re-registers all existing listeners.