Backup cleanup
The main change is that the Backup will immediately deserialize the backup Operation in the deserialize
method. There is no intermediate Data required.

This is done by adding an extra method to the ObjectDataInput.readDataAsObject.

Also the need to publish the actual BackupOperation in the Backup isn't needed any longer (publishing
was needed for diagnostics purposes).