Fix partition replica version handling when backups are reordered
When backups are reordered and a backup with a version greater than (localVersion + 1)
is received it's applied but partition replica versions are not updated
. After that point, until replica is repaired, succeeding stale backups cannot be detected
anymore. This breaks monotonicity of backups.

Now, partition replica versions are always updated when (newVersion > localVersion) but
partition is marked as dirty if (newVersion > localVersion + 1). Anti-entropy system will take
this dirty flag into account too.