If client can't connect, diagnostics not started
The problem is that in the start method, the call to clusterService.start blocks till a connection
is made. Once the connection is made, the diagnostics is started.

This is a problem when connection can't be made. THis pr fixed that by moving the diagnostics start
in front of the clusterService.start