Fixes in Operation.waitTimeout update on retrying
In case of retrying a BlockingOperation, the waitTimeout is made smaller so that
a timed blocking operation eventually times out.

However there are some problems:
- it can happen that a positive timeout, becomes negative due to subtraction. A negiative
timeout is considered to be an infinite timeout. So this can cause problems

- it can happen that a blocking call doesn't wait as long as specified because of subtracking
the elapsed time, the call timeout is subtracted. This is a problem if the elapsed time is
smaller than the call timeout. This can happen because the OperationParker periodically sweeps
through all operations to expire them. This could happen 1 second after the operation
get executed, but it will still get a 60 second (default timeout) being deducted from the
remaining timeout.