Retry the invocation only if managed to deregister
Avoids a race between two retrying threads which could result in a leak inside
the Invocation Registry.

If the operation was never registered, every thread is allowed to retry it. This
is because invocation can fail even before registration and then deregistration
always fails. Allowing retrying to all threads seems safe because there is no
risk of concurrent retrying of an operation that didn't even make it to registration.

The race is still there that can lead to multiple concurrent
operation retries. Only one retry invocation will be found in
the registry; others will be unaccounted for.