Reset hot restart data before join
Joiner impls check if the node needs to reset its hot restart. If that it is the case, the node resets its hot restart data to be able to join to the cluster.

This change pollutes join mechanism because of hot restart mechanics. However, without it, users have to clear hot restart data manually if they start excluded nodes after the cluster is started.