Invocation lastHeartbeatMillis is reset
In case of a retrying operation, an old heartbeat could still be stored
on the invocation. If the invocation monitor sees this; it currently doesn't
understand that this from an old invocation; it will just conclude that the
invocation has not received any recent heartbeats.

This pr fixes that by resetting the lastHeartbeat when the invocation is made.