Allow declarative Cache Writer/Loader configuration when the Writer/Loader factory in not on client classpath
Previously we forced users to provide a factory have it on classpath of clients.
This is unnecessary as clients obviously do not use Writer/Loader.

It also makes declarative configuration simpler to use. Instead of forcing
users to implement a factory it's not possible to directly specify a class to use.

So instead of
<cache-loader-factory class-name="com.foo.MyCacheLoaderFactory" />
it's now possible to use:
<cache-loader class-name="com.package.YourMapLoaderImpl" />