Make eventual assertion in PartitionServiceSafetyCheckTest
We don't check in the partition service if added / removed member is a lite member. If it is the case, the migration flow will be executed anyway without doing any work. As it is a cheap operation and we don't want to pollute related methods with lite member checks, we prefer to tune the test assertion accordingly.