Fixes #5453
Notify SENDER first - reason why this is so important:
Someone may do map.get(other_nodes_key) and when it finishes do map.loadAll
The problem is that map.get may finish earlier than then overall loading on the SENDER due to the fact
that the LoadStatusOperation may first reach the node that did map.get and not the SENDER.
The SENDER will be then in the LOADING status, thus the loadAll call will be ignored.
It happens only if all LoadAllOperation finish before the sendKeyLoadCompleted is started (test case, little data).