Fix near cache lite member test
The test first sends invalidation events to the lite member and then fills the near cache. If the invalidation events arrive after the near cache is filled, the test fails. This fix first fills the near cache and waits for the invalidation events to arrive before filling the near cache again.

Fixes : https://github.com/hazelcast/hazelcast/issues/8840