Fix semaphore & lock dead member cleanup mechanisms
Cleanup operations which are executed after a member is removed
should handle retry cases too. A partition can be promoted and/or migrated
when a member is removed. As a result of that, cleanup operation can be
rejected with a retry response.

With this change, cleanup operations are submitted to local member
via invocation mechanism.

Also simplified Semaphore, ILock and CountDownLatch split brain tests.