Further fix related to #9273 and #9271
 Related pull requests
 https://github.com/hazelcast/hazelcast/pull/9271
 https://github.com/hazelcast/hazelcast/pull/9273

In #9271, connections with resources are destroyed. I have converted
it so that every that endpoints are removed. The purpose of heartbeat
was to clean resources of a client when heartbeat is timeout in the
first place.

With the fix above, the addition of findLiveConnection which added in
if and only if its endpoint destroyed. When enpoint destroyed listeners
are removed. This means there can not be any listener task that its connection
is not live but it is trying to send events to client.