Fix wrong estimation due to wrong encoding engine.
When using an estimator to constantly add new elements and not checking the estimate
then encoding engine is not upgraded to the Dense, which leads to two problems.
a. the default engine keeps consuming more and more resourses and it is foundamentaly very hungry and slow
b. upon calling estimate at the very end, the converstion gives very wrong results

The problem with the delayed conversion between the encoders, is that the actual test on whether we should
upgrade, takes place on the estimate() call, rather than the add() call.

This patch, moves the check to the add() call instead, which slows down a bit, the fast path, but in the long term
this shouldn't affect it a lot.