Split ClockTest into a QuickTest and SlowTest
ClockTest seems to crash our CI builds quite often. The test method
test_whenConfiguringClockOffset_thenSystemOffsetClockIsCreated() takes
over 3 seconds to complete. The SystemClockChangeTest which uses the
same classloader magic takes up to 3 minutes (it's a NightlyTest).

This commit splits up the ClockTest into a really quick unit test of all
single components and skips the classloading magic for the singleton
part of the Clock implementation. This should fix the immediate issue
with our builds, but needs further investigation of the root cause.