Fix IMap query runner primary migration detection mechanism
Current in-flight migration detection mechanism is racy. `partitionStateVersion`
observed at the beginning of query execution can be the already updated `partitionStateVersion`.
So, any change of `partitionStateVersion` may not be noticed.

There's no ordering between `commitMigration()` / `rollbackMigration()` and update of `partitionStateVersion`, because we can apply promotion commits and source migration commits in batches to increase availability. That's why `commitMigration()` / `rollbackMigration()` and update of `partitionStateVersion` can be executed concurrently.

Instead of comparing `partitionStateVersion` and counting in-flight migrations, Map service now uses two counters; one for started migrations and one for completed migrations. We always increment counters on each migration.
`CountingMigrationAwareService` exposes a migration stamp when requested and a method to check whether stamps is still valid. If stamp is valid, that means there's no ongoing and/or completed change in primary partition ownership on that member.