Fixed classloader leakage through ThreadLocal<>
Replaced use of ThreadLocal<Random> subclass in ThreadLocalRandom (used
to override initialValue()) with a method that explicitly gets the
thread local value and sets it if it's null.

This avoids some nasty classloader issues that show up particularly in
web applications, where this anonymous subclass is holding a reference
to the entire web app classloader, which is not released until the
thread exits.

This basically replaces ThreadLocalRandom as backport of JDK 7 with a
simple holder class for leakage safe ThreadLocal<Random> and
ThreadLocal<SecureRandom> fields.

Made UuidUtil use of the new ThreadLocalRandom instead of its own
ThreadLocal fields.