Synchronize default config values and fix xml generator
Here we move the old config compatibility checker to the test code since it is no longer used in production. We also expand on this compatibility checker to include as many configuration as possible and use it to check if the default values in the XML config, the Java config and hazelcast-default.xml are the same. The XML generator tests were already using this compatibility checker so expanding it uncovered some missing config values. The changes :
- fix the "default" value in the hazelcast-config.xsd for the networkConfig/joinConfig/multicastConfig/loopbackModeEnabled from true to false
- fix documentation for the some default values
- add fluent API for generating XML
- add missing tags to xml generator