Fix wrong scheduled task count caused by backup replicas (#9704)
* Fix wrong scheduled task count caused by backup replicas

When scheduling tasks on random partitions, the getAllScheduled call
returns twice as many futures from the expected result. The reason
seems to be the backup replicas which are also returned, even though they are
just stashed placeholders having no future scheduled. To address this, we introduced
a flag, marking the main copy of the task as master upon initial scheduling or promotions
and using that when building the list of tasks to return to the API call