Do not require our Distributed version of functional interfaces (#296)
Currently, equivalents of java.util.function.* interfaces in Distributed
class are required as parameters to DistributedStream. Technically, we only
need to ensure, that the provided implementations are serializable and not
to require implementing our interface.

The Distributed versions are to be used as overloaded parameter type, so
that lambda expressions need not be casted.

Additionally, we actually try to serialize given objects to check if they are
really serializable to catch problems early, as the later stack trace will be
much more cryptic.