Server side ClientListener fix
ClientListener behaviour tested to confirm that it is fired
when client connection disconnected/connected to member.
It is listening only local connections.
Dummy client does not cause any event on members that it is
not connected.
A client reauthenticating to same connection also firing
clientDisconnected event. Fixed with this commit. SEVERE log
An endpoint already exists is deleted. It is printed when
client is reauthenticating over same connection. And it
is expected behaviour.