Fix lock eviction scheduling on migration when lock expired
Previously on lock migration the lock was scheduled for eviction with difference between the current cluster clock and the expiration time. If the migration lasted long enough and lock has already expired it could schedule expiration with a negative value. This would cause the lock to be expired after a period which is equal to the time elapsed from the actual expiration time - the more the migration is delayed, the more will the lease be extended.

Added some javadoc for lock classes, fixed some generics and simplified a couple of if-statements.