Fixes testListenersSmartRoutingTerminateRandomNode
test was assuming member added event is enough to wait.
It could be the case that member is added to cluster but client
is not yet added the listener back.
Note: test was failing when random node is owner node.

As a fix we are waiting for connection added event specfically.

fixes https://github.com/hazelcast/hazelcast/issues/9803