Refactor of Jet collectors into a new type called Reducer
The `Reducer` is a generalisation of j.u.s `Collector` and
models the terminal operation of a stream, and has the
responsbility to build the DAG and execute it.

The new interface also allows a distinction between j.u.s.
Collectors which can be used downstream, and Jet-specific
constructs such as `DistributedCollectors.toIMap()`

Fixes #309