Backport of #6702 and #7162
Invocation timeout handling is fixed. # of times an
invocation made is used to be count with 1 second sleeps. Since
invocation itself can take time, user configured timeout was
used to be violated. It is now checked by comparing
System.currentTimeMillis and starting time.
Half backport of #6702

Reauthenticate as owner to an old connection when owner dead
When client searching for a new owner connection, if it come accross
already established one, it will resend authentication
request to change the owner connection info on the cluster.
backport of #7162