Fix re-execution of completed task when killing the owner node (#9995)
* Fix re-execution of completed task when killing the owner node

When a task is submitted to a key owner, and the task completes, we replace the runnable
with the result in the internal ring buffer. However, there is no backup-operation to sync
the result with the backup nodes. If the node shuts down gracefully then the replication
will properly handle it, but when the node is killed, the backup nodes still have the task
in their ring-buffer rather than the result, thus, we re-execute it.

Fix #9965