[BACKPORT] Fix racy scheduled executor, syncState when task is cancelled. (#10034)
Cancelling an in-flight task (not handling interrupts properly), results in a cancel task state, which is racy with the syncState when the task finally completes execution. This is only documenting itself when replication happens, where the result of the task is used as decision mechanism to re-schedule it or not on the new partition owner.

The patch is gating the result assignment to check whether it was set before.
If so, the new result is ignored.

Fix #9984
(cherry picked from commit 4c7b5b71f984f4d27ab8ebc1fbead42ddb971d40)