Fix spurious crash in ClientExceptionFactory
The ExceptionUtilsTest in jet crashed spuriously with

```java.lang.ArrayIndexOutOfBoundsException
    at java.lang.System.arraycopy(Native Method)
    at com.hazelcast.client.impl.protocol.ClientExceptionFactory.createExceptionMessage(ClientExceptionFactory.java:708)
    ...
```

We couldn't trace down the reason (because after adding some log messages it
did not crash). Seems like the stacktraces or causes of the Throwables changed
under our hands, however unlikely or unbelievable that sounds.

I've just rewritten the stuff to call the `getStackTrace` and `getCause` just
once, and the problem disappeared (at least does not fail after 1000
repeats, while before it failed after 40 trials or so).