[BACKPORT] Fix scheduled task not getting rescheduled after dead member (#10092)
This patch is fixing the following logic defect.
When a migration starts, upon replication of a given task, the Future associated
with the task is cancelled. If a migration fails (rollback) or for any other reason
the partition gets promoted once again on the same member, the task is trying to reschedule itself.
However, the algorithm responsible for rescheduling it, it first checks if the task is completed or
running, by checking the result and the future variables respectively. If either of them is found non-null
then the task will not be re-scheduled. To fix, when the task is cancelled, we nullify the Future associated
and reset the isTaskOwner flag, both will set again when and if the task gets another chance to run.

Fix #10060

(cherry picked from commit 1aaa774ab821dbaaaa422a222a5347383a87119a)