Add discovery SPI implementation for AWS
Previously the AWS module supported only discovering members through the TcpIpJoiner mechanism. This was configured through the AwsConfig class. This mechanism is still available for backwards compatibility and here we will add support for the discovery SPI. By doing so, the user can now configure AWS using the old AwsConfig or the new discovery config and usage in Hazelcast core is simplified for future functionalities based on discovery SPI.

Currently the discovery SPI implementation will transform the discovery configuration into the old AwsConfig and reuse the existing code. In the future both the deprecated TcpIpJoiner and the AwsConfig can be removed.