Fixes async backup accumulation.
The backpressure mechanism for async backups has been modified. Instead of having a single counter
per partition, there is a single global counter that triggers periodic syncs. The problem with the
counter per partition, is that it is still possible to get a huge number of backup operations using
the default configuration. By default 271 partitions, 100 async per partition, giving 27,100 pending backups.

With the new approach this is simplified to at most 100 async backups before a sync is given.

Also a logging warning is given to increase the backup timeout if back pressure is enabled and a low (default) timeout value is given.

Just like the old approach, randomization is done to prevent lock-step problems. So if configured with a
sync window of 100, sync delay is randomized between 75 and 125 async backups.