Fixes client invocationfuture callback backpressure
THe idea is that for every callback, an invocationslot is taken.
And for every completion of a callback, an invocation slot is released.

So if there are many pending invocations, there will be few invocation
slots. Wich will reduce the pressure, and give the callbacks time to
complete instead of getting overloaded.