Removed contended counter in OperationRunnerImpl
All OperationRunnerImpl are contending on an executed operations
atomiclong. This causes contention for no good reason.

So there is now only 1 contented case and thes are the ad-hoc runners.

As a bonus we now also get metrics on number of operations executed
per generic operation runner and for the ad hoc operation runner.