Moved deserialization of response from response thread
Before this change, when a response-packet is received, it deserialized
to a (Normal)Response on the response thread. And then handed over to
the calling thread to do the real deserialization if needed. The problem
is that this puts a lot of pressure on the response thread and it becomes
a scalability bottleneck; especially with larger responses.

This pr changes this behavior so that instead of deserializing the response,
we poke in the bytes without needing to deserialize and handover the Packet
to the invocation future which will do these 2 consequtive deserializations
packet -> NormalResponse -> payload.

In a follow up PR, this intermediate deserialization of NormalResponse
will be fully removed. Removing one unwanted walk over the data + removing
temporary litter.