This pr contains a few things
1: there is some undefined behavior in the NonBlockingThreadingModel where on shutdown
the io threads array elements get nilled. This has been changed by just setting the whole
 array to null. This way the array remains effectively immutable.

2: a new diagnostics networking plugin that currently renders information about io threads.
 This should help to spot networking problems. This plugin works for both server and client.

2: modifications to the non blocking io threads so it contains the total number of bytes,
frames etc. This is done with a zero overhead on the fast path. So instead of having contented
counters, periodically a task is written to each read/write handler to publish information to
 its owning io thread. This is disabled by default and only works when debug metrics level is enabled.

3: the non blocking readers/writers have a configurable mechanism for returning 'load'. So byte based,
frame based, handle-method-count etc. The old behavior was always handle method count, but this doesnt'
seem to give the best results. Byte based seems to be more reliable. Currently these properties are only
configurable through a 'hidden' system property. In the 3.9 release we need to pick a better one and use
that be default. But at least we can now experiment with the different policies.