NativeMemoryData.toByteArray prevention
This is the first step where NativeMemoryData.toByteArray is prevented.
This method is being used to copy data from offheap into a byte array
and then it is being copied into the final byte array, e.g a response
for members, or a client response.

This creates litter and unwanted copying.

In this PR the ByteArrayDataOutputStream is optimized and
the ClientMessage.