Inject HazelcastInstance in backup entry processor If the supplied entry procesor extended AbstractEntryProcessor and implemented HazelcastInstanceAware, the backup processor would never get the hazelcast instance injected and the process method could fail with NPE.
The EntryBackupProcessorImpl now implements HazelcastInstanceAware
which checks if the outer object implements it as well and calls
the method to set the instance.

This changes the serialVersionUID of the EntryBackupProcessorImpl
which would cause it to fail deserialization when sent from a member
containing the fix to a member which doesn't contain the fix. This is
why we set the serialVersionUID to the value calculated for the
EntryBackupProcessorImpl without the fix.

Fixes : hazelcast#10083