NormalResponse optimizations
Skips the NormalResponse creation in case of responses for responses
that don't require a ack from a backup.

Most importantly: it also skips an intermediate copy of the heap/native-data
into a temporary serialization buffer and then into the packet. With this change
the data is immediately copied into the packet-bytes.

This pr is nativememorydata friendly; it doesn't cause unwanted call to toByteArray.